+++ app/controllers/template_options_controller.rb
@@ -7,6 +7,8 @@ class TemplateOptionsController < ApplicationController
 
   after_action :verify_authorized
 
+  DEFAULT_FUNDER_ID = Rails.application.config.default_funder_id
+
   # GET /template_options  (AJAX)
   # Collect all of the templates available for the org+funder combination
   # rubocop:disable Metrics/AbcSize, Metrics/MethodLength
@@ -17,7 +19,7 @@ class TemplateOptionsController < ApplicationController
     authorize Template.new, :template_options?
 
     org = org_from_params(params_in: { org_id: org_hash.to_json }) if org_hash.present?
-    funder = Org.find(Rails.application.config.default_funder_id)
+    funder = Org.find(DEFAULT_FUNDER_ID)
 
     @templates = []
 
@@ -53,7 +55,9 @@ class TemplateOptionsController < ApplicationController
     end
     @templates = @templates.flatten
 
-    @templates = sort_templates(@templates, org)
+    respond_to do |format|
+      format.json { render json: build_templates_payload(@templates) }
+    end
   end
   # rubocop:enable Metrics/AbcSize, Metrics/MethodLength
   # rubocop:enable Metrics/CyclomaticComplexity, Metrics/PerceivedComplexity
@@ -69,39 +73,21 @@ class TemplateOptionsController < ApplicationController
     %i[id name url language abbreviation ror fundref weight score]
   end
 
-  # Orders templates for dropdown:
-  # 1. Org templates
-  # 2. Priority funder templates (Simplified, Default)
-  # 3. Remaining funder templates
-  def sort_templates(templates, org)
-    # Get the funder templates with the Priority templates ordered in front
-    sorted_funder_templates = sort_funder_templates(templates)
-
-    # If an org was selected and the selected org is not the default funder
-    if org&.id && org.id != Rails.application.config.default_funder_id
-      # Return the templates with the org templates ordered in front
-      org_templates = templates.select { |t| t.org_id == org.id }
-      org_templates + sorted_funder_templates
-    # else, either no org or default_funder was selected
-    else
-      # All of the templates are funder templates
-      sorted_funder_templates
-    end
-  end
-
-  # Orders funder templates:
-  # 1. Priority (Simplified, Default)
-  # 2. Remaining funder templates
-  def sort_funder_templates(templates)
-    funder_templates = templates.select { |t| t.org_id == Rails.application.config.default_funder_id }
-
-    # Identify priority templates
-    # NOTE: This will have to be updated if the `simplified` template is ever renamed
+  # Returns a JSON payload with the following structure:
+  # - templates:
+  #   - org_templates: [All non-funder templates]
+  #   - priority_templates: [The simplified and default funder templates]
+  #   - other_templates: [All non-priority funder templates]
+  # - total_templates: Count of all templates
+  def build_templates_payload(templates)
+    funder_templates = templates.select { |t| t.org_id == DEFAULT_FUNDER_ID }
     simplified = funder_templates.find { |t| t.title.start_with?('Alliance Simplified Template') }
-    default   = funder_templates.find(&:is_default)
-    priority  = [simplified, default].compact
+    default    = funder_templates.find(&:is_default)
+    priority_templates = [simplified, default].compact
 
-    # Return priority first, then remaining funder templates
-    priority + (funder_templates - priority)
+    { templates: { org_templates: templates - funder_templates,
+                   priority_templates: priority_templates,
+                   other_templates: funder_templates - priority_templates },
+      total_templates: templates.size }
   end
 end